Troubleshooting

+++ cPWP/bamsToPWP.h
@@ -15,6 +15,6 @@
 int runANGSDforReadCounts (std::string bamlist, std::string angsdPrefix, std::string nThreads="25", std::string angsdOutputLog="angsdOutPut.log");
 int convertANGSDcountsToBinary(std::string angsdPrefix, std::string binaryOutputFileName, int numIndividuals, int readDepthMax);
 unsigned char readCountToUnsignedChar(int individual, std::string allele, std::vector<std::string> countsVector);
-int calcPWPfromBinaryFile (std::string binaryFile, int numLoci, const int numIndividuals, std::string outFile);
+int calcPWPfromBinaryFile (std::string binaryFile, unsigned long long int numLoci, const int numIndividuals, std::string outFile);
 
 #endif /* defined(__cPWP__readsToPWP__) */
